requirements.txt
# Scraper dependencies
tweepy>=4.14.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
python-dateutil>=2.8.2

# Transcriber dependencies
openai>=1.0.0
# Optional: local whisper (if you don't want to use OpenAI API)
# openai-whisper>=20231117
# torch>=2.0.0
# transformers>=4.36.0
# accelerate>=0.25.0
